+++ examples/hello-world-plugin/install.sh
@@ -1,0 +1,3 @@
+#!/bin/sh
+set -eu
+mv "$out/bin/hello-world-openclaw" "$out/bin/hello-world"